requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
openai==1.55.3
python-dotenv==1.0.1
requests==2.31.0
PyAudio==0.2.14
gunicorn==21.2.0
flask_socketio==5.3.6
SpeechRecognition==3.10.3
playsound==1.3.0
elevenlabs==1.1.2
gevent==21.12.0
